always display the maximum value in the data series
in other words, always print the exact height of the y axis

I have a feeling there's a more natural way to make the maximum y-axis
label reflect the maximum value in the range. But this works for now and
has the added utility of labeling the first point where the data series
actually hit the max value.